TEmos ate teste 4 correto

+++ src/homeaway/EatingClass.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 
 public class EatingClass extends ServicesClassAbstract implements Eating, Serializable {
 
-    public EatingClass(long latitude, long longitude, Double price, Double value, String serviceName){
+    public EatingClass(long latitude, long longitude, double price, int value, String serviceName){
         super(latitude,longitude,price,value,serviceName);
     }
